pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hubspotapilib"
version = "0.0.1"
authors = [
    { name="Gibran Dar", email="gibrandar@gmail.com" }
]
description = "A basic Hubspot API Library"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
dependencies = [
    "requests",
    "hubspot_api_client",
    "python-dotenv"
]

[project.urls]
"Homepage" = "https://github.com/GibranDar/hubspot-api-lib"
"Bug Tracker" = "https://github.com/GibranDar/hubspot-api-lib/issues"

